v1.144:
HIGHLIGHTS:
- resolved inability to use read-only depth or stencil and keeping the other subresource writable
- added ability to separate depth and stencil in a barrier
- introduced "PlaneBits" concept, which is also needed for video

BREAKING CHANGES:
- "attachmentContentType" renamed to "planes"
- "ClearDesc::planes" must be set to "PlaneBits::COLOR" explicitly for color only clears
- no changes in barriers unless depth and stencil separation needed
- removed "ResourceViewBits", use extended "Texture1DViewType" ("Texture2DViewType") instead
- no more API changes planned \O/

DETAILS:
- NRI: introduced "PlaneBits"
- NRI: extended "Texture1DViewType" and "Texture2DViewType" to include read-only depth and/or stencil
- NRI: removed "ResourceViewBits" (there was not a "resource" concept), replaced with extended view types
- NRI: "AttachmentContentType" replaced with "PlaneBits"
- NRI: added "PlaneBits" to "TextureBarrierDesc"
- D3D11/D3D12/VK: improvements around "ClearAttachments"
- VK: "ClearAttachments" behavior matched D3D
- Helper: added "planes" to "TextureUploadDesc"
- Validation: improved "ClearAttachments" validation
- Validation: added detection of a mismatched D3D/VK behavior related to read-only depth-stencil usage

+++ Source/D3D12/SharedD3D12.cpp
@@ -6,7 +6,7 @@
 
 using namespace nri;
 
-constexpr std::array<D3D12_COMMAND_LIST_TYPE, (uint32_t)CommandQueueType::MAX_NUM> COMMAND_LIST_TYPES = {
+constexpr std::array<D3D12_COMMAND_LIST_TYPE, (size_t)CommandQueueType::MAX_NUM> COMMAND_LIST_TYPES = {
     D3D12_COMMAND_LIST_TYPE_DIRECT,  // GRAPHICS,
     D3D12_COMMAND_LIST_TYPE_COMPUTE, // COMPUTE,
     D3D12_COMMAND_LIST_TYPE_COPY,    // COPY,
@@ -17,7 +17,7 @@ D3D12_COMMAND_LIST_TYPE nri::GetCommandListType(CommandQueueType commandQueueTyp
     return COMMAND_LIST_TYPES[(size_t)commandQueueType];
 }
 
-constexpr std::array<D3D12_RESOURCE_DIMENSION, (uint32_t)TextureType::MAX_NUM> RESOURCE_DIMENSIONS = {
+constexpr std::array<D3D12_RESOURCE_DIMENSION, (size_t)TextureType::MAX_NUM> RESOURCE_DIMENSIONS = {
     D3D12_RESOURCE_DIMENSION_TEXTURE1D, // TEXTURE_1D,
     D3D12_RESOURCE_DIMENSION_TEXTURE2D, // TEXTURE_2D,
     D3D12_RESOURCE_DIMENSION_TEXTURE3D, // TEXTURE_3D,
@@ -27,7 +27,7 @@ D3D12_RESOURCE_DIMENSION nri::GetResourceDimension(TextureType textureType) {
     return RESOURCE_DIMENSIONS[(size_t)textureType];
 }
 
-constexpr std::array<D3D12_DESCRIPTOR_RANGE_TYPE, (uint32_t)DescriptorType::MAX_NUM> DESCRIPTOR_RANGE_TYPES = {
+constexpr std::array<D3D12_DESCRIPTOR_RANGE_TYPE, (size_t)DescriptorType::MAX_NUM> DESCRIPTOR_RANGE_TYPES = {
     D3D12_DESCRIPTOR_RANGE_TYPE_SAMPLER, // SAMPLER
     D3D12_DESCRIPTOR_RANGE_TYPE_CBV,     // CONSTANT_BUFFER
     D3D12_DESCRIPTOR_RANGE_TYPE_SRV,     // TEXTURE
@@ -43,7 +43,7 @@ D3D12_DESCRIPTOR_RANGE_TYPE nri::GetDescriptorRangesType(DescriptorType descript
     return DESCRIPTOR_RANGE_TYPES[(size_t)descriptorType];
 }
 
-constexpr std::array<D3D12_PRIMITIVE_TOPOLOGY_TYPE, (uint32_t)Topology::MAX_NUM> PRIMITIVE_TOPOLOGY_TYPES = {
+constexpr std::array<D3D12_PRIMITIVE_TOPOLOGY_TYPE, (size_t)Topology::MAX_NUM> PRIMITIVE_TOPOLOGY_TYPES = {
     D3D12_PRIMITIVE_TOPOLOGY_TYPE_POINT,    // POINT_LIST
     D3D12_PRIMITIVE_TOPOLOGY_TYPE_LINE,     // LINE_LIST
     D3D12_PRIMITIVE_TOPOLOGY_TYPE_LINE,     // LINE_STRIP
@@ -79,7 +79,7 @@ D3D_PRIMITIVE_TOPOLOGY nri::GetPrimitiveTopology(Topology topology, uint8_t tess
         return PRIMITIVE_TOPOLOGIES[(size_t)topology];
 }
 
-constexpr std::array<D3D12_FILL_MODE, (uint32_t)FillMode::MAX_NUM> FILL_MODES = {
+constexpr std::array<D3D12_FILL_MODE, (size_t)FillMode::MAX_NUM> FILL_MODES = {
     D3D12_FILL_MODE_SOLID,    // SOLID
     D3D12_FILL_MODE_WIREFRAME // WIREFRAME
 };
@@ -88,7 +88,7 @@ D3D12_FILL_MODE nri::GetFillMode(FillMode fillMode) {
     return FILL_MODES[(size_t)fillMode];
 }
 
-constexpr std::array<D3D12_CULL_MODE, (uint32_t)CullMode::MAX_NUM> CULL_MODES = {
+constexpr std::array<D3D12_CULL_MODE, (size_t)CullMode::MAX_NUM> CULL_MODES = {
     D3D12_CULL_MODE_NONE,  // NONE
     D3D12_CULL_MODE_FRONT, // FRONT
     D3D12_CULL_MODE_BACK   // BACK
@@ -98,7 +98,7 @@ D3D12_CULL_MODE nri::GetCullMode(CullMode cullMode) {
     return CULL_MODES[(size_t)cullMode];
 }
 
-constexpr std::array<D3D12_COMPARISON_FUNC, (uint32_t)CompareFunc::MAX_NUM> COMPARISON_FUNCS = {
+constexpr std::array<D3D12_COMPARISON_FUNC, (size_t)CompareFunc::MAX_NUM> COMPARISON_FUNCS = {
 #ifdef NRI_USE_AGILITY_SDK
     D3D12_COMPARISON_FUNC_NONE, // NONE
 #else
@@ -118,7 +118,7 @@ D3D12_COMPARISON_FUNC nri::GetComparisonFunc(CompareFunc compareFunc) {
     return COMPARISON_FUNCS[(size_t)compareFunc];
 }
 
-constexpr std::array<D3D12_STENCIL_OP, (uint32_t)StencilFunc::MAX_NUM> STENCIL_OPS = {
+constexpr std::array<D3D12_STENCIL_OP, (size_t)StencilFunc::MAX_NUM> STENCIL_OPS = {
     D3D12_STENCIL_OP_KEEP,     // KEEP
     D3D12_STENCIL_OP_ZERO,     // ZERO
     D3D12_STENCIL_OP_REPLACE,  // REPLACE
@@ -133,7 +133,7 @@ D3D12_STENCIL_OP nri::GetStencilOp(StencilFunc stencilFunc) {
     return STENCIL_OPS[(size_t)stencilFunc];
 }
 
-constexpr std::array<D3D12_LOGIC_OP, (uint32_t)LogicFunc::MAX_NUM> LOGIC_OPS = {
+constexpr std::array<D3D12_LOGIC_OP, (size_t)LogicFunc::MAX_NUM> LOGIC_OPS = {
     D3D12_LOGIC_OP_NOOP,          // NONE
     D3D12_LOGIC_OP_CLEAR,         // CLEAR
     D3D12_LOGIC_OP_AND,           // AND
@@ -156,7 +156,7 @@ D3D12_LOGIC_OP nri::GetLogicOp(LogicFunc logicFunc) {
     return LOGIC_OPS[(size_t)logicFunc];
 }
 
-constexpr std::array<D3D12_BLEND, (uint32_t)BlendFactor::MAX_NUM> BLENDS = {
+constexpr std::array<D3D12_BLEND, (size_t)BlendFactor::MAX_NUM> BLENDS = {
     D3D12_BLEND_ZERO,             // ZERO
     D3D12_BLEND_ONE,              // ONE
     D3D12_BLEND_SRC_COLOR,        // SRC_COLOR
@@ -188,7 +188,7 @@ D3D12_BLEND nri::GetBlend(BlendFactor blendFactor) {
     return BLENDS[(size_t)blendFactor];
 }
 
-constexpr std::array<D3D12_BLEND_OP, (uint32_t)BlendFunc::MAX_NUM> BLEND_OPS = {
+constexpr std::array<D3D12_BLEND_OP, (size_t)BlendFunc::MAX_NUM> BLEND_OPS = {
     D3D12_BLEND_OP_ADD,          // ADD
     D3D12_BLEND_OP_SUBTRACT,     // SUBTRACT
     D3D12_BLEND_OP_REV_SUBTRACT, // REVERSE_SUBTRACT
@@ -200,7 +200,7 @@ D3D12_BLEND_OP nri::GetBlendOp(BlendFunc blendFunc) {
     return BLEND_OPS[(size_t)blendFunc];
 }
 
-constexpr std::array<D3D12_TEXTURE_ADDRESS_MODE, (uint32_t)AddressMode::MAX_NUM> TEXTURE_ADDRESS_MODES = {
+constexpr std::array<D3D12_TEXTURE_ADDRESS_MODE, (size_t)AddressMode::MAX_NUM> TEXTURE_ADDRESS_MODES = {
     D3D12_TEXTURE_ADDRESS_MODE_WRAP,   // REPEAT
     D3D12_TEXTURE_ADDRESS_MODE_MIRROR, // MIRRORED_REPEAT
     D3D12_TEXTURE_ADDRESS_MODE_CLAMP,  // CLAMP_TO_EDGE
@@ -211,7 +211,7 @@ D3D12_TEXTURE_ADDRESS_MODE nri::GetAddressMode(AddressMode addressMode) {
     return TEXTURE_ADDRESS_MODES[(size_t)addressMode];
 }
 
-constexpr std::array<D3D12_HEAP_TYPE, (uint32_t)MemoryLocation::MAX_NUM> HEAP_TYPES = {
+constexpr std::array<D3D12_HEAP_TYPE, (size_t)MemoryLocation::MAX_NUM> HEAP_TYPES = {
     D3D12_HEAP_TYPE_DEFAULT, // DEVICE
 #ifdef NRI_USE_AGILITY_SDK
     D3D12_HEAP_TYPE_GPU_UPLOAD, // DEVICE_UPLOAD (Prerequisite: D3D12_FEATURE_D3D12_OPTIONS16)
@@ -226,7 +226,7 @@ D3D12_HEAP_TYPE nri::GetHeapType(MemoryLocation memoryLocation) {
     return HEAP_TYPES[(size_t)memoryLocation];
 }
 
-constexpr std::array<D3D12_SHADING_RATE, (uint32_t)ShadingRate::MAX_NUM> SHADING_RATES = {
+constexpr std::array<D3D12_SHADING_RATE, (size_t)ShadingRate::MAX_NUM> SHADING_RATES = {
     D3D12_SHADING_RATE_1X1, // _1x1,
     D3D12_SHADING_RATE_1X2, // _1x2,
     D3D12_SHADING_RATE_2X1, // _2x1,
@@ -240,7 +240,7 @@ D3D12_SHADING_RATE nri::GetShadingRate(ShadingRate shadingRate) {
     return SHADING_RATES[(size_t)shadingRate];
 }
 
-constexpr std::array<D3D12_SHADING_RATE_COMBINER, (uint32_t)ShadingRate::MAX_NUM> SHADING_RATE_COMBINERS = {
+constexpr std::array<D3D12_SHADING_RATE_COMBINER, (size_t)ShadingRate::MAX_NUM> SHADING_RATE_COMBINERS = {
     D3D12_SHADING_RATE_COMBINER_OVERRIDE,    // REPLACE,
     D3D12_SHADING_RATE_COMBINER_PASSTHROUGH, // KEEP,
     D3D12_SHADING_RATE_COMBINER_MIN,         // MIN,
